# public/openapi.yaml
openapi: 3.1.0
info:
  title: DBackup API
  version: 3.4.0
  description: |
    REST API for DBackup - a self-hosted database backup automation platform with encryption, compression, and smart retention.

    ## Authentication

    DBackup supports two authentication methods:

    - **Session Authentication** - Used automatically when logged in via the web UI. Session cookies are sent with each request.
    - **API Key Authentication** - For scripts, CI/CD pipelines, and external integrations. Create an API key under **Access Management → API Keys**.

    > API keys do not inherit SuperAdmin privileges. Only explicitly assigned permissions are available.

    ## Rate Limits

    | Request Type | Default Limit |
    |---|---|
    | Authentication (login, etc.) | 5/min per IP |
    | GET requests | 100/min per IP |
    | POST / PUT / DELETE | 20/min per IP |

    Rate limits are configurable in Settings → Rate Limits.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
servers:
  - url: "{protocol}://{host}/api"
    description: DBackup Instance
    variables:
      protocol:
        default: https
        enum: [http, https]
      host:
        default: your-dbackup-instance.com
security:
  - BearerAuth: []
  - SessionAuth: []
...
paths:
...
  /storage/{id}/files/bulk:
    post:
      tags: [Storage]
      summary: Bulk Backup Action
      description: |
        Deletes, locks or unlocks several backups on a destination in a single request.

        A locked backup is never deleted - it is reported in `data.failed` instead. An
        incremental chain can be deleted as a whole, and members are removed newest-first
        so no snapshot is left depending on one that is gone. Selecting only part of a
        chain is still refused per file.

        **Required permission:** `storage:delete`
      operationId: bulkBackupAction
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, paths]
              properties:
                action:
                  type: string
                  enum: [delete, lock, unlock]
                paths:
                  type: array
                  items: { type: string }
                  minItems: 1
                  maxItems: 200
      responses:
        "200":
          description: Action applied, with per-backup outcomes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /storage/{id}/files:
    get:
      tags: [Storage Explorer]
      summary: List Backup Files
      description: Lists all backup files in a storage destination with enriched metadata.
      operationId: listBackupFiles
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - name: typeFilter
          in: query
          schema:
            type: string
          description: Filter by file type
      responses:
        "200":
          description: Backup file list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Storage Explorer]
      summary: Delete Backup File
      operationId: deleteBackupFile
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: File path to delete
                  example: "backups/mydb_2025-01-15.sql.gz.enc"
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API Key authentication. Create an API key under **Access Management → API Keys**.

        Format: `dbackup_xxxxxxxx`
    SessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie (browser login)
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  responses:
    Unauthorized:
      description: Missing, invalid, disabled, or expired credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Valid credentials but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Human-readable error message"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    BulkResponse:
      type: object
      description: |
        Outcome of an action applied to several records.

        `success` describes the request, not the records. It is `true` whenever the request
        was handled, including when every record failed - the per-record verdicts live in
        `data`. Only authentication, permission, validation and unexpected server errors
        produce `success: false` with a 4xx or 5xx status.
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: One-line summary, for example "8 of 10 jobs deleted".
          example: 8 of 10 jobs deleted
        data:
          type: object
          properties:
            succeeded:
              type: array
              description: Ids (or paths, for storage) that were processed.
              items: { type: string }
            failed:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                    description: Human-readable label, when one is known.
                  error:
                    type: string
                    description: Why this record was not processed.
...
    FileInfo:
      type: object
      properties:
        name:
          type: string
          example: "mydb_2025-01-15.sql.gz.enc"
        path:
          type: string
        size:
          type: integer
          description: Size in bytes
        lastModified:
          type: string
          format: date-time
        locked:
          type: boolean
          description: Whether the file is currently being written
...